token argument added, space comand moved to separated file
